+++ resources/META-INF/plugin.xml
@@ -398,7 +398,8 @@
 
         <!-- Project settings -->
         <projectService serviceImplementation="nl.hannahsten.texifyidea.run.latex.logtab.ui.LatexErrorTreeViewConfiguration"/>
-        <sdkType implementation="nl.hannahsten.texifyidea.settings.LatexSdk" />
+        <sdkType implementation="nl.hannahsten.texifyidea.settings.TexliveSdk" />
+        <sdkType implementation="nl.hannahsten.texifyidea.settings.DockerSdk" />
 
         <!-- Languages -->
         <lang.parserDefinition language="Latex" implementationClass="nl.hannahsten.texifyidea.LatexParserDefinition"/>
