+++ src/nl/hannahsten/texifyidea/documentation/LatexDocumentationProvider.kt
@@ -10,7 +10,7 @@ import nl.hannahsten.texifyidea.lang.Package.Companion.DEFAULT
 import nl.hannahsten.texifyidea.psi.BibtexEntry
 import nl.hannahsten.texifyidea.psi.BibtexId
 import nl.hannahsten.texifyidea.psi.LatexCommands
-import nl.hannahsten.texifyidea.run.latex.LatexDistribution
+import nl.hannahsten.texifyidea.settings.LatexSdk
 import nl.hannahsten.texifyidea.util.parentsOfType
 import nl.hannahsten.texifyidea.util.previousSiblingIgnoreWhitespace
 import java.io.IOException
@@ -119,7 +119,7 @@ class LatexDocumentationProvider : DocumentationProvider {
         val stream: InputStream
         try {
             // -M to avoid texdoc asking to choose from the list
-            val command = if (LatexDistribution.isTexliveAvailable) {
+            val command = if (LatexSdk.isTexliveAvailable) {
                 "texdoc -l -M $name"
             }
             else {
@@ -138,7 +138,7 @@ class LatexDocumentationProvider : DocumentationProvider {
             emptyList()
         }
         else {
-            if (LatexDistribution.isTexliveAvailable) {
+            if (LatexSdk.isTexliveAvailable) {
                 lines.map {
                     // Line consists of: name version path optional file description
                     it.split("\t")[2]
