+++ src/nl/hannahsten/texifyidea/run/latex/LatexRunConfiguration.kt
@@ -30,6 +30,7 @@ import nl.hannahsten.texifyidea.run.compiler.LatexCompiler.Format
 import nl.hannahsten.texifyidea.run.latex.logtab.LatexLogTabComponent
 import nl.hannahsten.texifyidea.run.latex.ui.LatexSettingsEditor
 import nl.hannahsten.texifyidea.run.linuxpdfviewer.PdfViewer
+import nl.hannahsten.texifyidea.settings.LatexSdk
 import nl.hannahsten.texifyidea.settings.TexifySettings
 import nl.hannahsten.texifyidea.util.allCommands
 import nl.hannahsten.texifyidea.util.files.commandsInFileSet
@@ -472,7 +473,7 @@ class LatexRunConfiguration constructor(
     }
 
     fun setDefaultDistribution() {
-        latexDistribution = LatexDistribution.defaultLatexDistribution
+        latexDistribution = LatexSdk.defaultLatexDistribution
     }
 
     /**
